+++ Tests/LinuxMain.swift
@@ -8,9 +8,11 @@ XCTMain([
     testCase(FileTests.allTests),
     testCase(ModuleTests.allTests),
     testCase(OffsetMapTests.allTests),
+    testCase(SourceKitObjectTests.allTests),
     testCase(SourceKitTests.allTests),
     testCase(StringTests.allTests),
     testCase(StructureTests.allTests),
+    testCase(SwiftDocKeyTests.allTests),
     testCase(SwiftDocsTests.allTests),
     testCase(SyntaxTests.allTests)
 ])