+++ Tests/SourceKittenFrameworkTests/CodeCompletionTests.swift
@@ -12,11 +12,11 @@ import XCTest
 
 class CodeCompletionTests: XCTestCase {
 
-    func testSimpleCodeCompletion() {
+    func testSimpleCodeCompletion() throws {
         let file = "\(NSUUID().uuidString).swift"
         let completionItems = CodeCompletionItem.parse(response:
-            Request.codeCompletionRequest(file: file, contents: "0.", offset: 2,
-                arguments: ["-c", file, "-sdk", sdkPath()]).send())
+            try Request.codeCompletionRequest(file: file, contents: "0.", offset: 2,
+                                               arguments: ["-c", file, "-sdk", sdkPath()]).send())
         compareJSONString(withFixtureNamed: "SimpleCodeCompletion",
                           jsonString: completionItems)
     }