+++ Tests/SourceKittenFrameworkTests/SwiftDocsTests.swift
@@ -15,10 +15,10 @@ func compareJSONString(withFixtureNamed name: String,
                        rootDirectory: String = fixturesDirectory,
                        file: StaticString = #file,
                        line: UInt = #line) {
-    #if os(Linux)
+#if os(Linux) && !swift(>=4.1)
     let jsonString = String(describing: jsonString).replacingOccurrences(of: rootDirectory, with: "")
     let actualContent = jsonString
-    #else
+#else
     // Strip out fixtures directory since it's dependent on the test machine's setup
     let escapedFixturesDirectory = rootDirectory.replacingOccurrences(of: "/", with: "\\/")
     let jsonString = String(describing: jsonString).replacingOccurrences(of: escapedFixturesDirectory, with: "")
@@ -28,13 +28,24 @@ func compareJSONString(withFixtureNamed name: String,
     let actualContent = absolutePathRegex.stringByReplacingMatches(in: jsonString, options: [],
                                                                    range: NSRange(location: 0, length: jsonString.bridge().length),
                                                                    withTemplate: "\"key\\.filepath\" : \"\",")
-    #endif
+#endif
 
     let expectedFile = File(path: versionedExpectedFilename(for: name))!
 
+    // Use if changes are introduced by changes in SourceKitten.
     let overwrite = false
     if overwrite && actualContent != expectedFile.contents {
         _ = try? actualContent.data(using: .utf8)?.write(to: URL(fileURLWithPath: expectedFile.path!), options: [])
+        return
+    }
+
+    // Use if changes are introduced by the new Swift version.
+    let appendFixturesForNewSwiftVersion = false
+    if appendFixturesForNewSwiftVersion && actualContent != expectedFile.contents,
+        var path = expectedFile.path, let index = path.index(of: "@"),
+        !path.hasSuffix("@\(buildingSwiftVersion).json") {
+        path.replaceSubrange(index..<path.endIndex, with: "@\(buildingSwiftVersion).json")
+        _ = try? actualContent.data(using: .utf8)?.write(to: URL(fileURLWithPath: path), options: [])
         return
     }
 
@@ -51,7 +62,7 @@ func compareJSONString(withFixtureNamed name: String,
 
     if jsonValue(actualContent) != jsonValue(expectedFile.contents) {
         XCTFail("output should match expected fixture", file: file, line: line)
-        print("actual:\n\(actualContent)\nexpected:\n\(expectedFile.contents)")
+        print(diff(original: expectedFile.contents, modified: actualContent))
     }
 }
 
@@ -62,20 +73,32 @@ private func compareDocs(withFixtureNamed name: String, file: StaticString = #fi
 }
 
 private func versionedExpectedFilename(for name: String) -> String {
-    #if swift(>=4.0)
-        let versions = ["swift-4.0", "swift-3.2", "swift-3.1"]
-    #elseif swift(>=3.2)
-        let versions = ["swift-3.2", "swift-3.1"]
-    #else // if swift(>=3.1)
-        let versions = ["swift-3.1"]
+    #if swift(>=4.2)
+        let versions = ["swift-4.2", "swift-4.1.50", "swift-4.1.2", "swift-4.1.1", "swift-4.1", "swift-4.0.3", "swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.1.50)
+        let versions = ["swift-4.1.50", "swift-4.1.2", "swift-4.1.1", "swift-4.1", "swift-4.0.3", "swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.1.2)
+        let versions = ["swift-4.1.2", "swift-4.1.1", "swift-4.1", "swift-4.0.3", "swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.1.1)
+        let versions = ["swift-4.1.1", "swift-4.1", "swift-4.0.3", "swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.1)
+        let versions = ["swift-4.1", "swift-4.0.3", "swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.0.3)
+        let versions = ["swift-4.0.3", "swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.0.2)
+        let versions = ["swift-4.0.2", "swift-4.0"]
+    #elseif swift(>=4.0)
+        let versions = ["swift-4.0"]
+    #else
+        fatalError("Swift 4.0 or later is required!")
     #endif
     #if os(Linux)
         let platforms = ["Linux", ""]
     #else
         let platforms = [""]
     #endif
-    for version in versions {
-        for platform in platforms {
+    for platform in platforms {
+        for version in versions {
             let versionedFilename = "\(fixturesDirectory)\(platform)\(name)@\(version).json"
             if FileManager.default.fileExists(atPath: versionedFilename) {
                 return versionedFilename
@@ -84,6 +107,56 @@ private func versionedExpectedFilename(for name: String) -> String {
     }
     return "\(fixturesDirectory)\(name).json"
 }
+
+private func diff(original: String, modified: String) -> String {
+    do {
+        let url = URL(fileURLWithPath: NSTemporaryDirectory()).appendingPathComponent("SwiftDocsTests-diff-\(NSUUID())")
+        try FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
+
+        try original.data(using: .utf8)?.write(to: url.appendingPathComponent("original.json"))
+        try modified.data(using: .utf8)?.write(to: url.appendingPathComponent("modified.json"))
+
+        let task = Process()
+        task.launchPath = "/usr/bin/env"
+        task.currentDirectoryPath = url.path
+        task.arguments = ["git", "diff", "original.json", "modified.json"]
+
+        let pipe = Pipe()
+        task.standardOutput = pipe
+        task.standardError = pipe
+
+        task.launch()
+
+        let file = pipe.fileHandleForReading
+        defer { file.closeFile() }
+
+        return String(data: file.readDataToEndOfFile(), encoding: .utf8) ?? ""
+    } catch {
+        return "\(error)"
+    }
+}
+
+private let buildingSwiftVersion: String = {
+    #if swift(>=4.2)
+        return "swift-4.2"
+    #elseif swift(>=4.1.50)
+        return "swift-4.1.50"
+    #elseif swift(>=4.1.2)
+        return "swift-4.1.2"
+    #elseif swift(>=4.1.1)
+        return "swift-4.1.1"
+    #elseif swift(>=4.1)
+        return "swift-4.1"
+    #elseif swift(>=4.0.3)
+        return "swift-4.0.3"
+    #elseif swift(>=4.0.2)
+        return "swift-4.0.2"
+    #elseif swift(>=4.0)
+        return "swift-4.0"
+    #else
+        fatalError("Swift 4.0 or later is required!")
+    #endif
+}()
 
 class SwiftDocsTests: XCTestCase {
 
@@ -130,9 +203,8 @@ extension SwiftDocsTests {
         return [
             ("testSubscript", testSubscript),
             ("testBicycle", testBicycle),
-            ("testExtension", testExtension)
-            // Fails on Linux
-            // ("testParseFullXMLDocs", testParseFullXMLDocs),
+            ("testExtension", testExtension),
+            ("testParseFullXMLDocs", testParseFullXMLDocs)
         ]
     }
 }